+++ lib/velocity_x.dart
@@ -1,6 +1,7 @@
 library velocity_x;
 
 export 'src/flutter/animated/animated_container.dart';
+export 'src/flutter/appbar.dart';
 export 'src/flutter/card.dart';
 export 'src/flutter/center.dart';
 export 'src/flutter/common/velocity_conditional.dart';
@@ -19,7 +20,9 @@ export 'src/flutter/padding.dart';
 export 'src/flutter/rich_text.dart';
 export 'src/flutter/scroll.dart';
 export 'src/flutter/selectable_text.dart';
+export 'src/flutter/shimmer.dart';
 export 'src/flutter/sizedbox.dart';
+export 'src/flutter/swiper.dart';
 export 'src/flutter/text.dart';
 export 'src/flutter/transform.dart';
 export 'src/flutter/widgets.dart';